Makes instruction pointer 32 bit to allow more than 64K of program memory. Fixed preamble parsing to auto-detect crossing 64k boundaries.

+++ cs/src/automata_runner.h
@@ -20,7 +20,7 @@ using std::vector;
 #include "nmranet/TractionClient.hxx"
 #include "executor/Timer.hxx"
 
-typedef uint16_t aut_offset_t;
+typedef unsigned aut_offset_t;
 typedef uint8_t insn_t;
 
 namespace nmranet { class Node; }
